requirements.txt
amqp==5.3.1
APScheduler==3.6.3
asgiref==3.8.1
billiard==4.2.1
cachetools==4.2.2
celery==5.5.3
certifi==2025.4.26
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2.3
djangorestframework==3.16.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==13.15
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tornado==6.1
tzdata==2025.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.13
